backend/requirements.txt
flatbuffers==25.2.10
libclang==18.1.1
mpmath==1.3.0
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
Pillow==10.0.1
protobuf==4.25.8
pyasn1==0.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sympy==1.14.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
